# openapi.yaml
openapi: 3.1.0
info:
  title: Familieassistenten API
  description: |
    REST API for Familieassistenten — en familie-husassistent for RPi5.
    Dekker måltidsplanlegging, handleliste, husarbeid, kalender, inventory,
    og LLM-chat. Alle feilsvar bruker RFC 7807 (application/problem+json).
  version: 1.1.0
  license:
    name: Privat
servers:
  - url: http://localhost:3000
    description: Lokal utvikling
  - url: http://familieassistenten.local
    description: RPi5 produksjon
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    Problem:
      type: object
      description: RFC 7807 Problem Details
      properties:
        type: { type: string, format: uri, example: "about:blank" }
        title: { type: string, example: "Bad Request" }
        status: { type: integer, example: 400 }
        detail: { type: string }
        instance: { type: string }
        errors:
          type: array
          description: Field-level validation errors
          items:
            type: object
            properties:
              path: { type: string }
              message: { type: string }
...
    MealSlot:
      type: object
      properties:
        id: { type: integer }
        weekYear: { type: string, example: "2026-W15" }
        dayOfWeek: { type: integer, minimum: 0, maximum: 6 }
        dayName: { type: string }
        mealType: { type: string, example: "middag" }
        recipeId: { type: integer, nullable: true }
        status: { type: string, enum: [planned, cooked, skipped, away] }
        recipe: { $ref: "#/components/schemas/Recipe" }
...
    Recipe:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        category: { type: string, enum: [rask, comfort, helg] }
        prepTime: { type: string }
        servings: { type: integer }
        ingredients:
          type: array
          items: { $ref: "#/components/schemas/RecipeIngredient" }
    RecipeIngredient:
      type: object
      properties:
        id: { type: integer }
        productKey: { type: string, nullable: true }
        name: { type: string }
        qty: { type: number }
        unit: { type: string }
        optional: { type: boolean }
...
  responses:
    BadRequest:
      description: Validation or malformed request
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
    NotFound:
      description: Resource or route not found
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
security:
  - bearerAuth: []
paths:
...
  /api/meals/swap:
    put:
      summary: Replace a meal slot with a different recipe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [dayOfWeek, recipeId]
              properties:
                weekYear: { type: string }
                dayOfWeek: { type: integer, minimum: 0, maximum: 6 }
                recipeId: { type: integer }
      responses:
        '200':
          description: Swap done
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  mealPlan:
                    type: array
                    items: { $ref: "#/components/schemas/MealSlot" }
        '400': { $ref: "#/components/responses/BadRequest" }
...
  /api/recipes/{id}:
    get:
      summary: Get a single recipe
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Recipe
          content:
            application/json:
              schema:
                type: object
                properties:
                  recipe: { $ref: "#/components/schemas/Recipe" }
        '404': { $ref: "#/components/responses/NotFound" }
...